requirements.txt
openai>=1.0.0
arize-phoenix-otel>=0.14.0
openinference-instrumentation-openai>=0.1.40
opentelemetry-api>=1.20.0
opentelemetry-sdk>=1.20.0
flask>=2.3.0
flask-cors>=4.0.0
gunicorn>=21.2.0